refactor: Infra 모듈 태그, 역량 키워드 도메인 분리(#46)
